Create README.md
Petronas Talk was very interesting honestly! the talk was very  informative and helpful to me as a new student to gain a picture about the operations of one the biggest companies in malaysia ,
The session also inspired me to think critically about how businesses in the energy sector can align their goals with sustainability and innovation. As a student interested in this field, I was motivated to explore further how I can contribute to such efforts, whether through research, technology, or policy advocacy.