Add files via upload
Reflection
This program has really been instrumental in providing me with the skills and the knowledge required to identify diverse challenges related to Technology and Computer science fields. It enabled me to determine the major problems that face our society and affect its technological growth. And then use a structured approach to gather information that will lead to the best solution that can address the problem. The design of the solution will highly depend on my creativity and the approach that I follow in information’s Elicitation process. Nevertheless this project has equipped me with multiple skills, designing the prototype for example was really joyful, also this program helped me alot when it comes to team working ,since I did everything after a series of discussions with my creative teammates . Lastly I will never forget this experience since it was really inspiring and impactful on me in order to provide help to some people in our society who are in a dire need for solutions like this.